+++ src/environments/environment.development.ts
@@ -1,6 +1,6 @@
 export const environment = {
   layout: 'test',
-  version: '21.0.9',
+  version: '21.0.10',
   defaultPage: 'color-picker',
   static: 'https://ngnest.com/static'
 };