+++ src/environments/environment.ts
@@ -1,6 +1,6 @@
 export const environment = {
   layout: 'index',
-  version: '21.0.9',
+  version: '21.0.10',
   defaultPage: 'docs',
   static: 'https://ngnest.com/static'
 };